refactor: api 요청 함수 주석 추가
