+++ src/apis/auth/index.ts
@@ -1,17 +1,19 @@
 import AuthAxios from '@/apis/@core/authInstance';
 import Axios from '@/apis/@core/instance';
 import {
-  AvailabilityType,
-  CategoryResponse,
   CheckAvailabilityResponse,
+  checkAvailabilityType,
   ILoginResponse,
   IRefreshResponse,
   ISendVerifyCodeResponse,
   ISocialLoginResponse,
   LoginDataType,
+  ResetPasswordType,
   SocialLoginTokenType,
+  VerifyCodeType,
 } from '@/types/auth';
 
+// 로그인
 export async function login({ email, password }: LoginDataType) {
   const response = await Axios.post<ILoginResponse>('/api/auth/login', {
     email,
@@ -21,6 +23,7 @@ export async function login({ email, password }: LoginDataType) {
   return response.data;
 }
 
+// 로그아웃
 export async function logout(accessToken: string) {
   const response = await AuthAxios.post('/api/auth/logout', {
     accessToken,
@@ -29,12 +32,14 @@ export async function logout(accessToken: string) {
   return response.data;
 }
 
+// 토큰 갱신
 export async function refresh() {
   const response = await Axios.post<IRefreshResponse>('/api/auth/refresh');
 
   return response.data;
 }
 
+// 이메일 인증코드 전송
 export async function sendVerifyEmail(email: string) {
   const response = await Axios.post('/api/auth/send-verification-code', {
     email,
@@ -43,13 +48,8 @@ export async function sendVerifyEmail(email: string) {
   return response.data;
 }
 
-export async function sendVerifyCode({
-  email,
-  code,
-}: {
-  email: string;
-  code: string;
-}) {
+// 이메일 인증코드 확인
+export async function sendVerifyCode({ email, code }: VerifyCodeType) {
   const response = await Axios.post<ISendVerifyCodeResponse>(
     '/api/auth/verify-email',
     {
@@ -61,6 +61,7 @@ export async function sendVerifyCode({
   return response.data;
 }
 
+// 소셜 로그인
 export async function sendSocialLoginToken({
   code,
   provider,
@@ -73,13 +74,11 @@ export async function sendSocialLoginToken({
   return response.data;
 }
 
+// 비밀번호 재설정
 export async function resetPassword({
   emailAuthToken,
   newPassword,
-}: {
-  emailAuthToken: string;
-  newPassword: string;
-}) {
+}: ResetPasswordType) {
   const response = await AuthAxios.post('/api/auth/reset-password', {
     emailAuthToken,
     newPassword,
@@ -88,13 +87,11 @@ export async function resetPassword({
   return response.data;
 }
 
+// 닉네임, 이메일 중복체크
 export const checkAvailability = async ({
   type,
   value,
-}: {
-  type: AvailabilityType;
-  value: string;
-}) => {
+}: checkAvailabilityType) => {
   const response = await Axios.get<CheckAvailabilityResponse>(
     `/api/auth/check-${type}?${type}=${value}`,
   );
@@ -102,12 +99,7 @@ export const checkAvailability = async ({
   return response.data;
 };
 
-export const fetchCategory = async () => {
-  const response = await Axios.get<CategoryResponse>(`/api/global/category`);
-
-  return response.data;
-};
-
+// 회원가입
 export const signUp = async (formData: FormData) => {
   const response = await Axios.post('/api/auth/register', formData);
